+++ Palette/Sources/Shared/Components/MessageBubble.swift
@@ -4,12 +4,12 @@ import FlowKit
 
 struct MessageBubble: View {
     let message: ChatMessageModel
+    let onImageTap: (UIImage) -> Void
     @State private var image: UIImage?
-    @State private var isFullScreenPresented = false
     @Flow var flow
-    let success_alert = Alert(title: "이미지 저장 완료!",
-                      message: "이미지가 사진첩에 저장되었어요!",
-                      dismissButton: .default("확인"))
+    let success_alert = Alert(title: Text("이미지 저장 완료!"),
+                              message: Text("이미지가 사진첩에 저장되었어요!"),
+                              dismissButton: .default(Text("확인")))
     
     private var formattedTime: String {
         let formatter = DateFormatter()
@@ -42,7 +42,9 @@ struct MessageBubble: View {
                     .frame(maxWidth: 200, maxHeight: 200)
                     .cornerRadius(13)
                     .onTapGesture {
-                        isFullScreenPresented = true
+                        if let image = self.image {
+                            onImageTap(image)
+                        }
                     }
                     .onAppear {
                         KingfisherManager.shared.retrieveImage(with: URL(string: message.message)!) { result in
@@ -60,15 +62,12 @@ struct MessageBubble: View {
                         Button(action: {
                             if let image = self.image {
                                 UIImageWriteToSavedPhotosAlbum(image, nil, nil, nil)
-                                flow.alert(success_alert, animated: true)
+//                                flow.alert(success_alert, animated: true)
                             }
                         }) {
                             Text("갤러리에 저장하기")
                             Image(systemName: "square.and.arrow.down")
                         }
-                    }
-                    .sheet(isPresented: $isFullScreenPresented) {
-                        FullScreenImageView(image: image, isPresented: $isFullScreenPresented)
                     }
             } else {
                 Text(message.message)
@@ -93,28 +92,5 @@ struct MessageBubble: View {
         Text(formattedTime)
             .font(.custom("SUIT-Regular", size: 11))
             .foregroundColor(.gray)
-    }
-}
-
-struct FullScreenImageView: View {
-    let image: UIImage?
-    @Binding var isPresented: Bool
-    
-    var body: some View {
-        NavigationView {
-            Group {
-                if let image = image {
-                    Image(uiImage: image)
-                        .resizable()
-                        .aspectRatio(contentMode: .fit)
-                        .edgesIgnoringSafeArea(.all)
-                } else {
-                    Text("이미지를 불러올 수 없습니다.")
-                }
-            }
-            .navigationBarItems(trailing: Button("닫기") {
-                isPresented = false
-            })
-        }
     }
 }
